feat: Add external link support for blog posts
Allows specifying an external URL for blog posts. If provided, clicking a blog post in the UI will navigate to this external link instead of the internal route. This enhances the ability to link to content hosted on other platforms.

Also includes necessary type and data structure updates to accommodate this new field.